### 2018-01-28 #552 @whedon
[PRE REVIEW]: ChemPy: A package useful for chemistry written in Python
**Submitting author:** @bjodah (<a href="http://orcid.org/0000-0003-0596-0222">Björn Dahlgren</a>)
**Repository:** <a href="https://github.com/bjodah/chempy" target ="_blank">github.com/bjodah/chempy</a>
**Version:** v0.6.3
