### 2018-01-28 a8f3ef9
More robust handling of equality of arrays with dimensions
