+++ joss-paper/paper.md
@@ -15,7 +15,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-ChemPy is a Python library which provides functions and classes for
+ChemPy is a Python library that provides functions and classes for
 solving chemistry related problems. It includes classes for
 representing substances, reactions, and systems of reactions. It also
 includes well established formulae from physical chemistry, as well as
@@ -24,17 +24,17 @@ in chemical kinetics. Last, but not the least, it collects
 parametrizations of chemical properties of substances from the
 literature.
 
-The class for substances are represented by name and optionally
-contain information on their composition, weight, charge etc., as
-well as how to pretty print them in e.g. LaTeX. Both the composition
-and pretty printing forms can be deduced by ChemPy's
+Substances are represented by a class that holds their names and, optionally,
+information on their composition, weight, charge etc., as
+well as how to pretty print them using LaTeX, HTML and unicode. Both the composition
+and stylistic representations can be deduced by ChemPy's
 parser. Reactions are represented through their stoichiometry and
 thermodynamic/kinetic parameters. If the stoichiometry of a reaction
 is unknown, ChemPy can balance it based on the composition of the
 substances. The classes for representing systems of reactions provide
 methods to analyze e.g. if there are disjoint sets of reactions, or if all
 are connected in the same network. The classes also offer
-a series of "checks" to be performed at construction, ensuring balanced
+a series of checks performed at initialization, ensuring balanced
 reactions with sane coefficients and consistent units.
 
 Systems of reactions can be represented as graphs, tables, systems of
@@ -46,10 +46,10 @@ numerically using pyodesys [@dahlgren_pyodesys_2018] and pyneqsys
 Thanks to the use of SymPy [@Meurer2017], a user can not only solve
 stoichiometry problems with a single unique solution, but also
 under-determined systems, where the answer then contains a free parameter.
-In fact, most equations and parametrizations in ChemPy support, in addition to
-NumPy [@vanderWalt2011] arrays, also symbolic input, as well as arrays
-with explicit units. The latter allows ChemPy to check that the correct
-dimensionality is used based on e.g. reaction order.
+In fact, most equations and parametrizations in ChemPy support---in addition to
+NumPy [@vanderWalt2011] arrays---also symbolic input, as well as arrays
+with explicit units. The latter allows ChemPy to check that e.g. the correct
+dimensionality is of a rate coefficient is used with respect to reaction order.
 
 # Features
 - Pretty printing of chemical formulae and reaction sets.
