### 2018-01-29 781b5d2
Add chempy.units.uniform
