### 2018-01-29 07029c6
Add rate_coeff method to Expr subclasses
