### 2018-01-29 efff6e3
Fix UnaryFunction.rate_coeff
