### 2018-01-29 66691c0
Release docs for v0.6.4
