### 2018-01-29 #552 @kyleniemeyer
@bryanwweber would you be able to review this submission?
