### 2018-01-29 d7e00cc
Update conda recipe
