+++ conda-recipe/meta.yaml
@@ -41,8 +41,6 @@ test:
     - {{ name }}.util
   requires:
     - pytest
-  commands:
-    - pytest --pyargs {{ name }}
 
 about:
   home: https://github.com/bjodah/{{ name }}
