### 2018-01-29 923b393
Minor fix to README and Docker-script
