+++ chempy/util/tests/test_graph.py
@@ -14,7 +14,7 @@ from ..graph import rsys2dot, rsys2graph
 from ..testing import requires
 
 try:
-    dot_missing = subprocess.call(['dot', '--version']) != 0
+    dot_missing = subprocess.call(['dot', '-?']) != 0
 except OSError:
     dot_missing = True
 
