### 2018-01-30 289e57a
Enhanced Jupyter JavaScript for ReactionSystem
