+++ chempy/kinetics/_native.py
@@ -107,7 +107,7 @@ def _get_subst_comp(rsys, odesys, comp_keys, skip_keys):
     return subst_comp
 
 
-def get_native(rsys, odesys, integrator, skip_keys=(0,), steady_state_root=False, conc_roots=None):
+def get_native(rsys, odesys, integrator, skip_keys=(0,), steady_state_root=False, conc_roots=None, clipping=None):
     comp_keys = Substance.composition_keys(rsys.substances.values(), skip_keys=skip_keys)
     if isinstance(odesys, PartiallySolvedSystem):
         init_conc = '&m_p[%d]' % (len(odesys.params) - len(odesys.original_dep))
@@ -133,6 +133,26 @@ def get_native(rsys, odesys, integrator, skip_keys=(0,), steady_state_root=False
             raise ValueError("integrator '%s' does not support roots." % integrator)
         if odesys.roots is not None:
             raise ValueError("roots already set")
+    if clipping:
+        if 'p_constructor' not in ns_extend:
+            ns_extend['p_constructor'] = []
+        ns_extend['p_constructor'] += [
+            'this->user_data = malloc(sizeof(double)*%(ny)d);' % dict(ny=odesys.ny)
+        ]
+        if 'p_destructor' not in ns_extend:
+            ns_extend['p_destructor'] = []
+        ns_extend['p_destructor'] += [
+            'free(this->user_data);'
+        ]
+        if 'p_includes' not in ns_extend:
+            ns_extend['p_includes'] = set()
+        ns_extend['p_includes'] |= {"<stdlib.h>"}
+        kw['namespace_override']['p_y_preprocessing'] = """
+        double * y = (double*)(this->user_data);
+        for (int i=0; i<%(ny)d; ++i){
+            y[i] = (y_[i] < 0) ? 0 : y_[i];
+        }
+        """ % dict(ny=odesys.ny)
     if steady_state_root:
         assert conc_roots is None
         kw['namespace_override']['p_nroots'] = ' return 1; '
