### 2018-02-01 #565 @whedon
[REVIEW]: ChemPy: A package useful for chemistry written in Python
**Submitting author:** @bjodah (<a href="http://orcid.org/0000-0003-0596-0222">Björn Dahlgren</a>)
**Repository:** <a href="https://github.com/bjodah/chempy" target ="_blank">github.com/bjodah/chempy</a>
**Version:** v0.6.3

Please avoid lengthy details of difficulties in the review thread. Instead, please create a new issue in the <a href="www.github.com/bjodah/chempy" target="_blank">target repository</a> and link to those issues (especially acceptance-blockers) in the review thread below. (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice versa will create corresponding breadcrumb trails in the link target.)

- [x] **Repository:** Is the source code for this software available at the <a target="_blank" href="www.github.com/bjodah/chempy">repository url</a>?
