### 2018-02-01 #565 @kyleniemeyer
here is the initial version of the compiled paper: https://github.com/openjournals/joss-papers/blob/joss.00552/joss.00552/10.21105.joss.00552.pdf
