### 2018-02-03 571431b
[WIP] clip_to_bounds for ODE system
