### 2018-02-06 fc44c06
Wrapp assert in np.all, return unit_registry in get_odesys extra
