### 2018-02-06 6a39f02
Release docs for v0.6.5
