### 2018-02-08 bjodah/chempy#80 @bryanwweber
Conda install instructions don't work on macOS
Hi! This is related to openjournals/joss-reviews#565

The conda installation instructions don't work for me on macOS (although they do on Ubuntu). The error I receive is

```bash
$ conda create -n chempy -c bjodah chempy pytest
Solving environment: failed

PackagesNotFoundError: The following packages are not available from current channels:

  - chempy
  - pyneqsys[version='>=0.5.1']
  - sym[version='>=0.3.1']
  - chempy
  - pyodesys[version='>=0.11.6']

Current channels:

  - https://conda.anaconda.org/bjodah/osx-64
  - https://conda.anaconda.org/bjodah/noarch
  - https://repo.continuum.io/pkgs/main/osx-64
  - https://repo.continuum.io/pkgs/main/noarch
  - https://repo.continuum.io/pkgs/free/osx-64
  - https://repo.continuum.io/pkgs/free/noarch
  - https://repo.continuum.io/pkgs/r/osx-64
  - https://repo.continuum.io/pkgs/r/noarch
  - https://repo.continuum.io/pkgs/pro/osx-64
  - https://repo.continuum.io/pkgs/pro/noarch
  - https://conda.anaconda.org/conda-forge/osx-64
  - https://conda.anaconda.org/conda-forge/noarch
```

You should list in the documentation/README which platforms are supported.
