### 2018-02-08 bjodah/chempy#81 @bryanwweber
Required and optional dependencies should be clarified in the documentation
This is related to openjournals/joss-reviews#565

Although the dependencies are documented in `setup.py`, the purpose of all of them is not very clear, and the format may not be familiar to users who have not seen a `setup.py` file before. The dependencies and the features they enable should be clarified in the documentation.
