### 2018-02-08 #565 @bryanwweber
@bjodah Since this comment is related more to the paper than to the software, I'll put it here. You should specify the target audience for the software - is it undergraduates doing homework problems, research scientists, government agencies, teachers, etc.
