### 2018-02-13 d598db9
Add support for dictionaries in chempy.units.unit_of
