### 2018-02-19 #565 @kyleniemeyer
@bryanwweber I noticed a few unchecked boxes above—are those issues related to the bjodah/chempy#80 and bjodah/chempy#81 issues you filed?
