### 2018-02-19 dede476
Differentiate str/unicode. Better docstring for ReactionSystem.from_string
