### 2018-02-19 #565 @bjodah
Sorry, I haven't addressed those yet. @bryanwweber pointed out that I don't provide conda packages for all platforms (only Linux). So I have been working on trying to support all (optional) dependencies of ChemPy for both OS X and Windows, but since many of them needs to be compiled with quite recent C++ features that endeavor quickly turned into a real rabbit hole.

I think I will just do what @bryanwweber suggested and clearly state that Linux is currently the "officially supported" platform, and then I'll continue to work on adding support for more platforms at a later stage. (I know people have gotten many things to work on OS X and windows using pip so it is doable but I don't have any computers with those systems for testing at the moment).
