### 2018-02-19 5d1e707
Elaborate on setup dependencies in README
