### 2018-02-19 #565 @kyleniemeyer
@bjodah ok—regarding the platform issue, that seems reasonable as long as you are clear. If you do limit the conda route to Linux, then it would be helpful to explain how the package can be installed/used on other platforms.
