### 2018-02-19 bjodah/chempy#81 @bjodah
I tried to elaborate on this [here](https://github.com/bjodah/chempy/commit/5d1e70754a25e1f1cfd1cb6a790b2887d4fc3d04), what do you think, is it too terse? I can of course be more detailed.
