### 2018-02-19 #565 @kyleniemeyer
@arfon FYI, the *Repository* link above seems to have some extra junk prepended on it, not sure if that has already been fixed in whedon: `https://github.com/openjournals/joss-reviews/issues/www.github.com/bjodah/chempy` instead of `www.github.com/bjodah/chempy`
