### 2018-02-19 bjodah/chempy#80 @bjodah
So I've spent some time trying to build conda packages for all the major platforms. It was harder than I anticipated but os x should be doable in the forseeable future (e.g. [working on it here](https://github.com/conda-forge/staged-recipes/pull/5110)). But until then I've updated the README to state that Linux is the officially supported platform.
