### 2018-02-19 83745c8
Add appveyor config and clarify install section of README
