### 2018-02-19 #565 @bryanwweber
@kyleniemeyer @bjodah The two about the "Statement of Need" are addressed by my comment on this issue: https://github.com/openjournals/joss-reviews/issues/565#issuecomment-363984324 The installation instructions are covered by the two issues in the ChemPy repository (and I'm satisfied with the response to them generally, which I'll mark on those two issues). The last one is the functionality testing, which is on me, I didn't get a chance to do that yet, will do so ASAP.
