### 2018-02-20 2c34562
Fix private printers in .util._expr
