### 2018-03-18 1e7066b
Try to utilize cvxopt in balance_stoichiometry
