+++ chempy/chemistry.py
@@ -1089,6 +1089,26 @@ class Equilibrium(Reaction):
         return candidate
 
 
+def _solve_balancing_ilp(A):
+    from cvxopt import matrix
+    from cvxopt.glpk import ilp
+
+    def as_matrix(mtx):
+        return matrix([[float(elem) for elem in row] for row in mtx.T.tolist()])
+
+    c_ = matrix([1]*A.shape[1])
+    A_ = as_matrix(A)
+    b_ = matrix([0.0]*A.shape[0])
+    G_ = matrix([[-1.0 if ri == ci else 0 for ci in range(A.shape[1])] for ri in range(A.shape[1])])
+    h_ = matrix([-1.0]*A.shape[1])
+
+    print("\n"+str(A)+"\n calling glpk.ilp") # DO-NOT-MERGE!
+    status, isol = ilp(c_, G_, h_, A_, b_, I=set(range(A.shape[1])))
+    if status not in 'optimal feasible'.split():
+        raise RuntimeError("cvxopt.glpk.ilp returned with status %s" % status)
+    return isol
+
+
 def balance_stoichiometry(reactants, products, substances=None,
                           substance_factory=Substance.from_formula,
                           parametric_symbols=None, underdetermined=True):
@@ -1140,8 +1160,8 @@ def balance_stoichiometry(reactants, products, substances=None,
 
     """
     from sympy import (
-        Matrix, gcd, zeros, linsolve, numbered_symbols, Wild, Symbol,
-        preorder_traversal as pre
+        MutableDenseMatrix, gcd, zeros, linsolve, numbered_symbols, Wild, Symbol,
+        Integer, Tuple, preorder_traversal as pre
     )
 
     _intersect = set.intersection(*map(set, (reactants, products)))
@@ -1175,7 +1195,7 @@ def balance_stoichiometry(reactants, products, substances=None,
     def _get(sk, ck):
         return substances[sk].composition.get(ck, 0) * (-1 if sk in reactants else 1)
 
-    A = Matrix([[_get(sk, ck) for sk in subst_keys] for ck in cks])
+    A = MutableDenseMatrix([[_get(sk, ck) for sk in subst_keys] for ck in cks])
     symbs = list(reversed([next(parametric_symbols) for _ in range(len(subst_keys))]))
     sol, = linsolve((A, zeros(len(cks), 1)), symbs)
     wi = Wild('wi', properties=[lambda k: not k.has(Symbol)])
@@ -1183,12 +1203,12 @@ def balance_stoichiometry(reactants, products, substances=None,
         lambda n: n.match(symbs[-1]/wi), pre(sol)) if m is not None])
     sol = sol.func(*[arg/cd for arg in sol.args])
 
-    def remove(sol, symb, remaining):
+    def remove(cont, symb, remaining):
         subsd = dict(zip(remaining/symb, remaining))
-        sol = sol.func(*[(arg/symb).expand().subs(subsd) for arg in sol.args])
-        if sol.has(symb):
+        cont = cont.func(*[(arg/symb).expand().subs(subsd) for arg in cont.args])
+        if cont.has(symb):
             raise ValueError("Bug, please report an issue at https://github.com/bjodah/chempy")
-        return sol
+        return cont
 
     done = False
     for idx, symb in enumerate(symbs):
@@ -1204,7 +1224,7 @@ def balance_stoichiometry(reactants, products, substances=None,
             break
         for expr in sol:
             if (expr/symb).is_number:
-                sol = remove(sol, symb, Matrix(symbs[idx+1:]))
+                sol = remove(sol, symb, MutableDenseMatrix(symbs[idx+1:]))
                 break
     for symb in symbs:
         cd = 1
[…]
