### 2018-03-18 3d44cf0
Use pulp instead of cvxopt
