### 2018-03-22 df44065
balance_stoichiometry: raise ValueError when components are missing
