### 2018-03-22 5720dd9
balance_stoichiometry: sanity checks before return
