+++ setup.py
@@ -98,7 +98,7 @@ _author, _author_email = open(_path_under_setup('AUTHORS'), 'rt').readline().spl
 
 extras_req = {
     'integrators': ['scipy>=0.16.1', 'pyodeint>=0.7.0', 'pycvodes>=0.8.3', 'pygslodeiv2>=0.6.1'],
-    'solvers': ['pykinsol', 'pulp'],
+    'solvers': ['pykinsol'],
     'native': ['pycompilation>=0.4.3', 'pycodeexport>=0.1.1', 'appdirs'],
     'docs': ['Sphinx', 'sphinx_rtd_theme', 'numpydoc'],
     'plotting': ['bokeh>=0.11.1', 'ipywidgets'],
@@ -121,7 +121,8 @@ setup_kwargs = dict(
     install_requires=[
         'numpy>1.8', 'scipy>=0.16.1', 'matplotlib>=1.5.3',
         'sympy>=1.1.1', 'quantities>=0.12.1', 'pyneqsys>=0.5.1',
-        'pyodesys>=0.11.6', 'pyparsing>=2.0.3', 'sym>=0.3.1', 'jupyter'
+        'pyodesys>=0.11.6', 'pyparsing>=2.0.3', 'sym>=0.3.1', 'jupyter',
+        'pulp>=1.6.8'
         # 'dot2tex>=2.9.0'
     ],
     extras_require=extras_req
