### 2018-03-23 76bff58
Print fraction coefficients properly
