+++ .circleci/config.yml
@@ -7,11 +7,11 @@ jobs:
     steps:
       - checkout
       - run: conda install --quiet gcc_linux-64 gxx_linux-64
-      - run: conda config --add channels conda-forge
-      - run: conda config --set show_channel_urls true
       - run: conda config --set always_yes yes
       - run: conda update --quiet conda
       - run: conda update --quiet python
+      - run: conda config --add channels conda-forge
+      - run: conda config --set show_channel_urls true
       - run: conda install --quiet pip pytest numpy scipy matplotlib sympy quantities pyparsing
       - run: pip install .
       - run: pytest -ra
