### 2018-03-24 #565 @bjodah
The second entry in the references in https://github.com/openjournals/joss-papers/blob/joss.00552/joss.00552/10.21105.joss.00552.pdf does not render properly. Any idea what's going wrong there?
