### 2018-03-24 #565 @kyleniemeyer
@bjodah the reference style we use avoids repeating the same author name multiple times in a row, so the line is just to indicate the same author as the reference before it
