### 2018-03-24 bjodah/chempy#81 @bryanwweber
This seems fine to me!
