### 2018-03-24 bjodah/chempy#80 @bryanwweber
OK this sounds good. Feel free to close if you want.
