### 2018-03-24 bjodah/chempy#88 @bryanwweber
examples/demo_kinetics.py throws TypeError
This is related to the review: https://github.com/openjournals/joss-reviews/issues/565

In trying to verify the functionality of the code, I ran into an error running examples/demo_kinetics.py. Running on Ubuntu 16.04, I installed ChemPy 0.6.5 from conda in a Python 3.6 environment, then cloned this repository, then

```bash
cd examples
python demo_kinetics.py
```

which results in

```python
Traceback (most recent call last):
  File "demo_kinetics.py", line 37, in <module>
    import argh
ModuleNotFoundError: No module named 'argh'

Traceback (most recent call last):
  File "demo_kinetics.py", line 39, in <module>
    main()
  File "demo_kinetics.py", line 25, in main
    expr = f(*args, **kwargs)
TypeError: pseudo_irrev() got an unexpected keyword argument 'exp'
```
