### 2018-03-24 #565 @bryanwweber
I'm satisfied that the package functions as specified, with the exception of an error in one of the demo scripts (linked above, issue on chempy repo). However, the paper appears to have not been updated with the target audience, as suggested in the review criteria, and as I noted in https://github.com/openjournals/joss-reviews/issues/565#issuecomment-363984324 Is there an updated version of the paper I missed?
