### 2018-03-26 f217caf
Better excpetion message in check_balance, fix check_positive
