+++ chempy/reactionsystem.py
@@ -299,10 +299,11 @@ class ReactionSystem(object):
                 else:
                     return True
         for rxn in self.rxns:
-            for net in rxn.composition_violation(self.substances):
+            for net, k in zip(*rxn.composition_violation(self.substances, composition_keys=True)):
                 if net != 0:
                     if throw:
-                        raise ValueError("Composition violation in %s" % str(rxn))
+                        raise ValueError("Composition violation (%s: %s) in %s" %
+                                         (k, net, rxn.string(with_param=False)))
                     else:
                         return False
         return True
