### 2018-03-27 2299d64
Better ValueError text, enhanced Reaction.rate_expr()
