+++ chempy/tests/test_chemistry.py
@@ -125,6 +125,13 @@ def test_Reaction_parsing():
     with pytest.raises(ValueError):
         Reaction.from_string('2 A -> B; 2e-3/hour', None)
 
+    r8 = Reaction.from_string('A -> B; "k"')
+    assert r8.rate_expr().args == None
+    assert r8.rate_expr().unique_keys == ('k',)
+    r9 = Reaction.from_string('A -> B; 42.0')
+    assert r9.rate_expr().args == [42.0]
+    assert r9.rate_expr().unique_keys == None
+
 
 @requires(parsing_library, units_library)
 def test_Substance__molar_mass():
