+++ setup.py
@@ -104,7 +104,7 @@ extras_req = {
     'plotting': ['bokeh>=0.11.1', 'ipywidgets'],
     'testing': ['pytest', 'pytest-cov', 'pytest-flakes', 'pytest-pep8']
 }
-extras_req['all'] = list(chain(extras_req.values())) + ['argh']
+extras_req['all'] = list(chain(extras_req.values()))
 
 setup_kwargs = dict(
     name=pkg_name,
@@ -122,7 +122,7 @@ setup_kwargs = dict(
         'numpy>1.8', 'scipy>=0.16.1', 'matplotlib>=1.5.3',
         'sympy>=1.1.1', 'quantities>=0.12.1', 'pyneqsys>=0.5.1',
         'pyodesys>=0.11.6', 'pyparsing>=2.0.3', 'sym>=0.3.1', 'jupyter',
-        'pulp>=1.6.8'
+        'pulp>=1.6.8',
         # 'dot2tex>=2.9.0'
     ],
     extras_require=extras_req
