### 2018-03-28 2bac5ee
Fix examples/demo_kinetics.py
