### 2018-03-28 bjodah/chempy#88 @bjodah
Thanks, fixed in gh-89. (argh wansn't even used so I dropped to import)
