### 2018-03-28 #565 @bjodah
(I updated the paper with  2 new references since a PR was merged in between with fixes that added PuLP as a dependency)
