### 2018-03-28 8917939
Release docs for v0.6.6
