+++ v0.6.6/_sources/index.rst.txt
@@ -1,0 +1,407 @@
+.. chempy documentation master file, created by
+   sphinx-quickstart on Wed Mar 28 18:06:49 2018.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to chempy's documentation!
+==================================
+
+.. toctree::
+   :maxdepth: 4
+   :caption: Contents:
+
+   chempy
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
+Overview
+========
+
+.. image:: http://hera.physchem.kth.se:9090/api/badges/bjodah/chempy/status.svg
+   :target: http://hera.physchem.kth.se:9090/bjodah/chempy
+   :alt: Build status
+.. image:: https://img.shields.io/pypi/v/chempy.svg
+   :target: https://pypi.python.org/pypi/chempy
+   :alt: PyPI version
+.. image:: https://img.shields.io/badge/python-2.7,3.5,3.6-blue.svg
+   :target: https://www.python.org/
+   :alt: Python version
+.. image:: https://zenodo.org/badge/8840/bjodah/chempy.svg
+   :target: https://zenodo.org/badge/latestdoi/8840/bjodah/chempy
+.. image:: https://img.shields.io/pypi/l/chempy.svg
+   :target: https://github.com/bjodah/chempy/blob/master/LICENSE
+   :alt: License
+.. image:: http://img.shields.io/badge/benchmarked%20by-asv-green.svg?style=flat
+   :target: http://hera.physchem.kth.se/~chempy/benchmarks
+   :alt: airspeedvelocity
+.. image:: http://hera.physchem.kth.se/~chempy/branches/master/htmlcov/coverage.svg
+   :target: http://hera.physchem.kth.se/~chempy/branches/master/htmlcov
+   :alt: coverage
+
+
+.. contents::
+
+
+About ChemPy
+------------
+`ChemPy <https://github.com/bjodah/chempy>`_ is a `Python <https://www.python.org>`_ package useful for
+chemistry (mainly physical/inorganic/analytical chemistry). Currently it includes:
+
+- Numerical integration routines for chemical kinetics (ODE solver front-end)
+- Integrated rate expressions (and convenience fitting routines)
+- Solver for equilibria (including multiphase systems)
+- Relations in physical chemistry:
+
+  - Debye-Hückel expressions
+  - Arrhenius & Eyring equation
+  - Einstein-Smoluchowski equation
+
+- Properties (pure python implementations from the litterature)
+
+  - water density as function of temperature
+  - water permittivity as function of temperature and pressure
+  - water diffusivity as function of temperature
+  - water viscosity as function of temperature
+  - sulfuric acid density as function of temperature & weight fraction H₂SO₄
+  - More to come... (and contributions are most welcome!)
+
+
+Documentation
+-------------
+The easiest way to get started is to have a look at the examples in this README,
+and also the jupyter notebooks_. In addition there is auto-generated API documentation
+for the latest `stable release here <https://bjodah.github.io/chempy/latest>`_
+(and `here are <http://hera.physchem.kth.se/~chempy/branches/master/html>`_ the API docs for the development version).
+
[…]
