### 2018-03-28 #565 @bjodah
I've tagged a new version (v0.6.6) and updated `joss-paper/codemeta.json` with the new version number and the new zenodo doi.
