### 2018-04-03 #565 @kyleniemeyer
Hi @bjodah, just a few comments on the paper:

- Could you add the city and country to your affiliation?
- Could you put commas around the "e.g." in the second, third, and last paragraphs?
- Reference `mitchell2011pulp` does not have correct bibliographic information (Citeseer is not a journal); I'm not sure what the correct info should be though... perhaps it is a conference paper, as suggested at https://www.semanticscholar.org/paper/PuLP%3A-A-Linear-Programming-Toolkit-for-Python-Mitchell-Consulting/24c9ad0d66f6a05ad41563a7dade60bff6f59106 ?
- in the last paragraph, I think you need a semicolon in the combined citation `[@mitchell2011pulp,@lougee2003common]`
