### 2018-04-04 4546118
Corrections to manuscript.
To address https://github.com/openjournals/joss-reviews/issues/565#issuecomment-378431281

- Comma after 'e.g.'
- Changed PuLP citation to a 'misc' entry (with some precedence in google scholar)
- Added city and country to 'affiliation' in codemeta.json
