+++ joss-paper/paper.bib
@@ -63,12 +63,11 @@
   pages = {87--90}
 }
 
-@article{mitchell2011pulp,
-  title={PuLP: a linear programming toolkit for python},
+@misc{mitchell2011pulp,
+  title={PuLP: a linear programming toolkit for python. The University of Auckland, Auckland, New Zealand},
   author={Mitchell, Stuart and OSullivan, Michael and Dunning, Iain},
-  journal={The University of Auckland, Auckland, New Zealand, http://www. optimization-online. org/DB\_FILE/2011/09/3178. pdf},
-  year={2011},
-  publisher={Citeseer}
+  url={http://www.optimization-online.org/DB_FILE/2011/09/3178.pdf}
+  year={2011}
 }
 
 @article{lougee2003common,
