### 2018-04-04 30f5674
Bibtex fix
