+++ joss-paper/paper.bib
@@ -66,7 +66,7 @@
 @misc{mitchell2011pulp,
   title={PuLP: a linear programming toolkit for python. The University of Auckland, Auckland, New Zealand},
   author={Mitchell, Stuart and OSullivan, Michael and Dunning, Iain},
-  url={http://www.optimization-online.org/DB_FILE/2011/09/3178.pdf}
+  url={http://www.optimization-online.org/DB_FILE/2011/09/3178.pdf},
   year={2011}
 }
 
