### 2018-04-04 #565 @bjodah
@kyleniemeyer thanks!
I (hopefully) fixed the points you raised. Let me know if you want me to change anything else.
