### 2018-04-04 4e0d2db
minor edits to paper
