### 2018-04-04 #565 @kyleniemeyer
@bjodah I made a few more in https://github.com/bjodah/chempy/pull/90, please merge that and then archive the repo to get a DOI
