### 2018-04-04 a8447b9
About to release v0.6.7
