+++ chempy/units.py
@@ -558,8 +558,8 @@ class Backend(object):
     >>> np.sum([1000*pq.metre/pq.kilometre, 1])
     1001.0
     >>> be_np = Backend(np)
-    >>> be_np.sum([[1000*pq.metre/pq.kilometre, 1], [3, 4]], axis=1)
-    array([ 2.,  7.])
+    >>> be_np.sum([[1000*pq.metre/pq.kilometre, 1], [3, 4]], axis=1)  # doctest: +NORMALIZE_WHITESPACE
+    array([2., 7.])
 
     """
 
