+++ scripts/generate_docs.sh
@@ -25,6 +25,7 @@ sed -i "s/$MATCH/$NEW/g" doc/conf.py
 sed -i "s/alabaster/sphinx_rtd_theme/g" doc/conf.py
 if [[ $NARGS -eq 3 ]]; then
 cat <<EOF>>doc/conf.py
+
 context = {
     'conf_py_path': '/doc/',
     'github_user': '$1',
