+++ setup.cfg
@@ -1,5 +1,6 @@
 [tool:pytest]
 norecursedirs = .* _* build dist conda-recipe scripts benchmarks doc deploy venv
+doctest_optionflags= NORMALIZE_WHITESPACE
 pep8maxlinelength=119
 pep8ignore =
     doc/conf.py ALL  # conf.py is a generated file
