### 2018-04-04 eba7ce4
Update codemeta.json with new doi for v0.6.7
- Addresses https://github.com/openjournals/joss-reviews/issues/565#issuecomment-378626546
