### 2018-04-04 #565 @bjodah
@kyleniemeyer thanks again. Merged, tagged & updated codemeta.json with new tag & DOI.
