### 2018-04-04 #565 @kyleniemeyer
@arfon this paper is now ready to accept.
