### 2018-04-04 0f51bff
Doctest annotations not working with pytest
Partially reverting a8447b962f6d1f0e09299e412a2446ffdf66c2e2.
